Restructured store folder → new filenames + created findings module (that extends crud)
