+++ src/store/index.js
@@ -2,8 +2,8 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 
 // for demo purposes only
-import database from './modules/database';
-import audits from './modules/audits/index.js';
+import database from './modules/database.js';
+import audits from './modules/audits.js';
 
 Vue.use(Vuex);
