+++ src/store/modules/findings.js
@@ -1,0 +1,17 @@
+import crud from './crud.js';
+
+export default {
+    amespaced: true,
+    actions: {
+        ...crud.actions,
+    },
+    getters: {
+        ...crud.getters,
+    },
+    state: {
+        ...crud.state,
+    },
+    mutations: {
+        ...crud.mutations,
+    }
+}
